fix msvc compiler warning about possible size_t/int value truncation
Warning reported when you attempt to implicitly convert between values of types
int & size_t as those types may be of different sized and int is signed while
size_t is not.

+++ source/bases/Console.c
@@ -53,11 +53,11 @@ int main(int argc, char **argv)
 #if PY_MAJOR_VERSION >= 3
     char fileName[MAXPATHLEN + 1];
     wchar_t **wargv, *wfileName;
-    int i, size;
+    size_t i, size;
 #else
     const char *fileName;
 #endif
-    int status;
+    size_t status;
 
     // initialize Python
 #ifndef CX_FREEZE_KEEP_PATH
